Split tests into two and and test for orgunit service endpoint
